Geometric models

GSplines.g0surface — Function

g0surface(hm::HMesh,d=3), g0surface(m::Mesh,d=3)

This function takes in input a mesh (or an half egde data structure of a mesh) and returns a g0 multipatch surface obtained using the Approximate Catmull-Clark scheme (ACC3)

Example

using G1Splines
m = offdata("cube.off")
g0 = g0surface(m)

GSplines.g1surface — Function

g1surface(hm::HMesh,S::String="CS-S")

This function takes in input a mesh in the half egde data structure and a string with the G1 solving strategy and will return an array composed by the patches constituing the surface.

By default, if two EVs are connected, the mesh is split. If the option check_ev = false, this does not happen, but the construction may be wrong.

Example

using G1Splines
m = offdata("cube.off")
g1 = g1surface(m)

GSplines.g1matrix — Function

Compute the matrix defining the G1 splines on the mesh hm with the knot distribution kn using quadratic glueing data.

It assumes that boundary vertices are regular.

GSplines.G1matrix — Function

This function creates the smoothing matrix to create a G1 smooth patch around extraordinary vertices of any valence

N–> Valence of the EV

S–> Solving strategy NCS-S (non circulant system+symm), NCS-AS(non circulant system+asymm), CS-AS(circulant system+asymm), CS-S(circulant system+symm)

T–>Type of mesh face The output matrix B has smoothing masks ordered as follow

Control points labelling for inner EVs Biquintic patch

    28 31 34 | 21 20 19
    29 32 35 | 24 23 22
    30 33 36 | 25 26 27
    -------------------       B=[1|2|3|...|34|35|36] columns OF B
    7  8  9  | 18 15 12
    4  5  6  | 17 14 11
    1  2  3  | 16 13 10

For regular vertices N=4 it will return ACC3 smoothing masks

GSplines.g1matrix_edge — Function

Compute the matrix defining the edge spline basis, in a compressed form.

GSplines.fake_edges — Function

This function fakes the half edge datas tructure of a mesh re-defining some edges to be boundaries from a given list of edges in input