Geometric models

G1Splines.g1surface — Function

g1surface(hm::HMesh,S::String="CS-S")

This function takes in input a mesh in the half egde data structure and a string with the G1 solving strategy and will return an array composed by the patches constituing the surface

G1Splines.g1matrix — Function

Compute the matrix defining the G1 splines on the mesh hm with the knot distribution kn using quadratic glueing data.

It assumes that boundary vertices are regular.

G1Splines.G1matrix — Function

This function creates the smoothing matrix to create a G1 smooth patch around extraordinary vertices of any valence

N–> Valence of the EV

S–> Solving strategy NCS-S (non circulant system+symm), NCS-AS(non circulant system+asymm), CS-AS(circulant system+asymm), CS-S(circulant system+symm)

T–>Type of mesh face The output matrix B has smoothing masks ordered as follow

Control points labelling for inner EVs Biquintic patch

    28 31 34 | 21 20 19
    29 32 35 | 24 23 22
    30 33 36 | 25 26 27
    -------------------       B=[1|2|3|...|34|35|36] columns OF B
    7  8  9  | 18 15 12
    4  5  6  | 17 14 11
    1  2  3  | 16 13 10

For regular vertices N=4 it will return ACC3 smoothing masks

G1Splines.g1matrix_edge — Function

Compute the matrix defining the edge spline basis, in a compressed form.