Geometric models

GSplines.g0surface — Function

g0surface(hm::{HMesh, Mesh}, d=3)

This function takes in input a mesh (or an half egde data structure of a mesh) and returns a (bicubic) g0 multipatch surface obtained using the Approximate Catmull-Clark scheme (ACC3). If the input degree is greater than 3, the resulting surface is degree elevated to the desired degree.

Example

using G1Splines
m = offdata("cube.off")
g0 = g0surface(m)

GSplines.g1surface — Function

g1surface(hm::HMesh,S::String="CS-S")

This function takes in input a mesh in the half egde data structure and a string with the G1 solving strategy and will return an array composed by the patches constituing the surface.

By default, if two EVs are connected, the mesh is split. If the option check_ev = false, this does not happen, but the construction may be wrong.

The input string contains the solving strategy for the construction of the G1 surface to be selected from the following four: "CS-S","CS-AS","NCS-S","NCS-AS". The default strategy is "CS-S".

Example

using G1Splines
m = offdata("cube.off")
g1 = g1surface(m)